Cleanup and simplified QCConnector tests
