+++ qc11-connector/build.gradle
@@ -12,3 +12,32 @@ dependencies {
     testImplementation "io.testerra:driver-ui-desktop:" + testerraTestVersion
     testImplementation "io.testerra:surefire-connector:" + testerraTestVersion
 }
+
+test {
+//    def profiles = [
+//            "restclient": "qc-restclient/src/test/resources/suite.xml",
+//    ]
+//
+//    def suiteFiles = []
+//    profiles.each { k, v ->
+//        if (project.hasProperty("" + k)) {
+////                def f = v
+//            suiteFiles << v
+//        }
+//    }
+
+    useTestNG() {
+        suites("src/test/resources/suite.xml")
+    }
+
+    testLogging {
+        outputs.upToDateWhen { false }
+        showStandardStreams = true
+    }
+
+    options {
+        systemProperties(System.getProperties())
+    }
+
+    ignoreFailures = true
+}
