+++ qc11-connector/src/test/java/eu/tsystems/mms/tic/testframework/qc11connector/test/AbstractQcTest.java
@@ -5,8 +5,9 @@
  * Riesaer Str. 5, 01129 Dresden
  * All rights reserved.
  */
-package eu.tsystems.mms.tic.testframework.qc11connector.test.abstracts;
+package eu.tsystems.mms.tic.testframework.qc11connector.test;
 
+import eu.tsystems.mms.tic.testframework.logging.Loggable;
 import eu.tsystems.mms.tic.testframework.qc11connector.constants.QCConstants;
 import eu.tsystems.mms.tic.testframework.qcrest.clients.QcRestClient;
 import eu.tsystems.mms.tic.testframework.qcrest.wrapper.TestRun;
@@ -14,6 +15,8 @@ import eu.tsystems.mms.tic.testframework.qcrest.wrapper.TestSet;
 import eu.tsystems.mms.tic.testframework.qcrest.wrapper.TestSetTest;
 import java.io.IOException;
 import java.util.List;
+
+import eu.tsystems.mms.tic.testframework.testing.TesterraTest;
 import org.testng.annotations.AfterSuite;
 
 /**
@@ -21,7 +24,7 @@ import org.testng.annotations.AfterSuite;
  *
  * @author sepr
  */
-public abstract class AbstractQcTest {
+public abstract class AbstractQcTest extends TesterraTest implements Loggable {
 
     /**
      * Remove ever growing stack of test runs in testsetundertest.
@@ -30,6 +33,7 @@ public abstract class AbstractQcTest {
      */
     @AfterSuite
     public void cleanUp() throws Exception {
+        log().info("Cleanup created testruns");
         String[] testsets = new String[]{QCConstants.QCSYNC3_TESTSET_NAME};
         for (String testSet : testsets) {
             TestSet ts = QcRestClient.getTestSet(testSet, QCConstants.QC_TESTSUNDERTEST_FOLDER);
@@ -42,6 +46,7 @@ public abstract class AbstractQcTest {
                         // (qcconnector Tests)
                         first = false;
                     } else {
+                        log().info("Remove testrun with ID " + run.getId());
                         QcRestClient.removeTestRun(run.getId());
                     }
                 }
