+++ qc11-connector/src/test/java/eu/tsystems/mms/tic/testframework/qc11connector/testsundertest/qcsync3/AttachmentUploadTests.java
@@ -8,17 +8,16 @@
 package eu.tsystems.mms.tic.testframework.qc11connector.testsundertest.qcsync3;
 
 import eu.tsystems.mms.tic.testframework.constants.TesterraProperties;
-import eu.tsystems.mms.tic.testframework.qc11connector.util.TestNgTestMethods;
 import eu.tsystems.mms.tic.testframework.qcrest.constants.QCProperties;
 import eu.tsystems.mms.tic.testframework.utils.UITestUtils;
 import eu.tsystems.mms.tic.testframework.webdrivermanager.WebDriverManager;
-import java.lang.reflect.Method;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import org.openqa.selenium.firefox.FirefoxProfile;
 import org.testng.Assert;
 import org.testng.annotations.AfterTest;
 import org.testng.annotations.BeforeMethod;
+import org.testng.annotations.Test;
+
+import java.lang.reflect.Method;
 
 /**
  * Test used to verify their screenshots or screencasts are uploaded.
@@ -37,11 +36,11 @@ public class AttachmentUploadTests {
         System.clearProperty(QCProperties.UPLOAD_SCREENSHOTS_PASSED);
         System.clearProperty(QCProperties.UPLOAD_SCREENSHOTS_FAILED);
         System.clearProperty(TesterraProperties.SCREENCASTER_ACTIVE);
-        if (!method.getName().contains("Screenshot")) {
-            final FirefoxProfile profile = new FirefoxProfile();
-            WebDriverManager.setGlobalExtraCapability(FirefoxDriver.PROFILE, profile);
-            WebDriver driver = WebDriverManager.getWebDriver();
-        }
+//        if (!method.getName().contains("Screenshot")) {
+//            final FirefoxProfile profile = new FirefoxProfile();
+//            WebDriverManager.setGlobalExtraCapability(FirefoxDriver.PROFILE, profile);
+//            WebDriver driver = WebDriverManager.getWebDriver();
+//        }
     }
 
     /**
@@ -56,7 +55,7 @@ public class AttachmentUploadTests {
     /**
      * Test using property qc.test.passed.upload.screenshots
      */
-    @org.testng.annotations.Test
+    @Test
     public void testScreenshotUploadAllPassed() {
         System.setProperty(QCProperties.UPLOAD_SCREENSHOTS_PASSED, "true");
         final WebDriver driver = WebDriverManager.getWebDriver();
@@ -68,7 +67,7 @@ public class AttachmentUploadTests {
     /**
      * Test using property qc.test.failed.upload.screenshots
      */
-    @org.testng.annotations.Test
+    @Test
     public void testScreenshotUploadAllFailed() {
         System.setProperty(QCProperties.UPLOAD_SCREENSHOTS_FAILED, "true");
         final WebDriver driver = WebDriverManager.getWebDriver();
@@ -76,39 +75,43 @@ public class AttachmentUploadTests {
         Assert.fail();
     }
 
+//    /**
+//     * Test using property qc.test.failed.upload.automatic.screenshot
+//     */
+//    @org.testng.annotations.Test
+//    public void testScreenshotUploadAutoFailed() {
+//        final WebDriver driver = WebDriverManager.getWebDriver();
+//        UITestUtils.takeScreenshot(driver, false);
+//        Assert.fail();
+//    }
+
     /**
-     * Test using property qc.test.failed.upload.automatic.screenshot
+     * Test using property qc.test.failed.upload.videos, default=true
      */
-    @org.testng.annotations.Test
-    public void testScreenshotUploadAutoFailed() {
+    @Test
+    public void testScreencastUploadFailed() {
         final WebDriver driver = WebDriverManager.getWebDriver();
         UITestUtils.takeScreenshot(driver, false);
         Assert.fail();
     }
 
-    /**
-     * Test using property qc.test.failed.upload.screencast
-     */
-    @org.testng.annotations.Test
-    public void testScreencastUploadFailed() {
-        TestNgTestMethods.failingWebDriverTest();
-        Assert.fail();
-    }
+//    /**
+//     * Test using property qc.test.failed.upload.videos
+//     */
+//    @Test
+//    public void testScreenshotUploadFallback() {
+//        final WebDriver driver = WebDriverManager.getWebDriver();
+//        UITestUtils.takeScreenshot(driver, false);
+//        Assert.fail();
+//    }
 
     /**
-     * Test using property qc.test.failed.upload.screencast
+     * Test using property qc.test.failed.upload.videos=false
      */
-    @org.testng.annotations.Test
-    public void testScreenshotUploadFallback() {
-        TestNgTestMethods.failingWebDriverTest();
-    }
-
-    /**
-     * Test using property qc.test.failed.upload.screencast
-     */
-    @org.testng.annotations.Test
+    @Test
     public void testAttachmentUploadNone() {
         System.setProperty(QCProperties.UPLOAD_VIDEOS, "false");
-        TestNgTestMethods.failingWebDriverTest();
+        WebDriverManager.getWebDriver();
+        Assert.fail();
     }
 }
